+++ examples/common/synchronous-client.py
@@ -66,7 +66,8 @@ client.connect()
 # The slave to query is specified in an optional parameter for each
 # individual request. This can be done by specifying the `unit` parameter
 # which defaults to `0x00`
-#---------------------------------------------------------------------------# 
+#---------------------------------------------------------------------------#
+log.debug("Reading Coils")
 rr = client.read_coils(1, 1, unit=0x01)
 
 #---------------------------------------------------------------------------# 
@@ -80,36 +81,51 @@ rr = client.read_coils(1, 1, unit=0x01)
 # blocks for the two sets, so a change to one is a change to the other.
 # Keep both of these cases in mind when testing as the following will
 # _only_ pass with the supplied async modbus server (script supplied).
-#---------------------------------------------------------------------------# 
+#---------------------------------------------------------------------------#
+log.debug("Write to a Coil and read back")
 rq = client.write_coil(0, True, unit=1)
 rr = client.read_coils(0, 1, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
 assert(rr.bits[0] == True)          # test the expected value
 
+log.debug("Write to multiple coils and read back- test 1")
 rq = client.write_coils(1, [True]*8, unit=1)
-rr = client.read_coils(1, 8, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
-assert(rr.bits == [True]*8)         # test the expected value
+rr = client.read_coils(1, 21, unit=1)
+assert(rr.function_code < 0x80)     # test that we are not an error
+resp = [True]*21
 
+# If the returned output quantity is not a multiple of eight,
+# the remaining bits in the final data byte will be padded with zeros
+# (toward the high order end of the byte).
+
+resp.extend([False]*3)
+assert(rr.bits == resp)         # test the expected value
+
+log.debug("Write to multiple coils and read back - test 2")
 rq = client.write_coils(1, [False]*8, unit=1)
 rr = client.read_coils(1, 8, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
 assert(rr.bits == [False]*8)         # test the expected value
 
 
+log.debug("Read discrete inputs")
 rr = client.read_discrete_inputs(0, 8, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
 
+log.debug("Write to a holding register and read back")
 rq = client.write_register(1, 10, unit=1)
 rr = client.read_holding_registers(1, 1, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
 assert(rr.registers[0] == 10)       # test the expected value
 
+log.debug("Write to multiple holding registers and read back")
 rq = client.write_registers(1, [10]*8, unit=1)
 rr = client.read_holding_registers(1, 8, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
 assert(rr.registers == [10]*8)      # test the expected value
 
+log.debug("Read input registers")
 rr = client.read_input_registers(1, 8, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error
 
@@ -119,6 +135,7 @@ arguments = {
     'write_address':   1,
     'write_registers': [20]*8,
 }
+log.debug("Read write registeres simulataneously")
 rq = client.readwrite_registers(unit=1, **arguments)
 rr = client.read_holding_registers(1, 8, unit=1)
 assert(rq.function_code < 0x80)     # test that we are not an error